+++ src/main/java/leets/bookmark/domain/bookmark/domain/service/BookmarkGetService.java
@@ -16,38 +16,11 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class BookmarkGetService implements BookmarkUseCase {
+public class BookmarkGetService {
 
     private final BookmarkRepository bookmarkRepository;
 
     public List<Bookmark> getBookmarksAllByMemo(String keyword) {
         return bookmarkRepository.findByMemoContaining(keyword);
-    }
-
-    @Override
-    public CommonResponse<List<BookmarkResponse>> getFilteredBookmarks(BookmarkFilterRequest request) {
-        List<Bookmark> bookmarks = bookmarkRepository.findAllWithFilter(request.categoryId(), request.tagNames());
-        List<BookmarkResponse> response = bookmarks.stream()
-            .map(BookmarkMapper::toResponse)
-            .toList();
-        return CommonResponse.createSuccess(BOOKMARK_FILTER_SUCCESS.getMessage(), response);
-    }
-
-    @Override
-    public CommonResponse<List<BookmarkResponse>> getByMemoContaining(String keyword) {
-        List<Bookmark> bookmarks = bookmarkRepository.findByMemoContaining(keyword);
-        List<BookmarkResponse> response = bookmarks.stream()
-            .map(BookmarkMapper::toResponse)
-            .toList();
-        return CommonResponse.createSuccess(BOOKMARK_MEMO_SEARCH_SUCCESS.getMessage(), response);
-    }
-
-    @Override
-    public CommonResponse<List<BookmarkResponse>> getAllBookmarks() {
-        List<Bookmark> bookmarks = bookmarkRepository.findAll();
-        List<BookmarkResponse> response = bookmarks.stream()
-            .map(BookmarkMapper::toResponse)
-            .toList();
-        return CommonResponse.createSuccess(BOOKMARK_SEARCH_SUCCESS.getMessage(), response);
     }
 }
